pull out repl utility functions + repl2 initCode / switching patterns works now
